refactor(domain): 替换 Agent 相关实体中的 Double 类型为 BigDecimal
- AgentDefinition 类中 temperature 字段由 Double 改为 BigDecimal，增强数值精度
- AgentTraceStep 类中 cost 字段由 Double 改为 BigDecimal，支持更精确的成本计算
- 删除消息模块中已废弃的消息批次及渠道评分相关服务与实现代码，简化代码库

+++ ydsz-agent/ydsz-agent-domain/src/main/java/com/njydsz/agent/domain/entity/AgentTraceStep.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * Agent 执行链路步骤（映射 ydsz_agt_trace_step 表）
  *
@@ -44,6 +46,6 @@ public class AgentTraceStep {
   /** 耗时（毫秒） */
   private Long durationMs;
 
-  /** Token 成本（ USD，精确到 6 位小数；非 LLM 调用步骤为 0） */
-  private Double cost;
+  /** Token 成本（USD，精确到 6 位小数；非 LLM 调用步骤为 0） */
+  private BigDecimal cost;
 }
